Fixed. Patch 1.1 Validation, redirect, button styles.

+++ src/main/java/web/dao/UserDaoImpl.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Repository;
 import web.model.User;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
@@ -36,6 +37,9 @@ public class UserDaoImpl implements UserDao {
     @Override
     public void deleteById(Long id) {
         User user = em.find(User.class, id);
+        if (user == null) {
+            throw new EntityNotFoundException("User with id " + id + " not found");
+        }
         em.remove(user);
     }
 }
